add optional delay

+++ cute_sway_recorder/main.py
@@ -1,18 +1,20 @@
 #!/usr/bin/env python3
 
-from pathlib import Path
 import signal
 import subprocess
 import sys
+from pathlib import Path
 from subprocess import DEVNULL
 from typing import Union
 
-from PySide6.QtCore import Qt
+from PySide6.QtCore import QParallelAnimationGroup, QTimer, Qt
+from PySide6.QtGui import QPalette
 from PySide6.QtWidgets import (
     QApplication,
     QGroupBox,
     QHBoxLayout,
     QLabel,
+    QMainWindow,
     QMessageBox,
     QPushButton,
     QStyle,
@@ -21,11 +23,16 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-from config_area import ConfigArea
 from common import SelectedArea, SelectedScreen
+from config_area import ConfigArea
+
+STATUS_DELAY = '<font color="indianred">Starting in {}...</font>'
+STATUS_RECORDING = '<font color="Red">RECORDING</font>'
+STATUS_NOT_RECORDING = "Not recording"
+STOP_RECORDING = "Stop recording"
 
 
-def start_recording(
+def wf_recorder(
     selection: Union[SelectedArea, SelectedScreen],
     file_dst,
     include_audio: bool = False,
@@ -50,61 +57,73 @@ def start_recording(
     return subprocess.Popen(params)
 
 
-class CuteRecorderQtApplication:
+class CuteRecorderQtApplication(QMainWindow):
     """
     Regular class handling state-y GUI stuff
     """
 
     def __init__(self):
+        QMainWindow.__init__(self)
         self.recorder_proc = None
 
-        self.app = QApplication(sys.argv)
-        self.app.setApplicationDisplayName("Cute Sway Recorder")
-        self.app.setDesktopFileName("cute-sway-recorder")
+        self.delay_timer = QTimer()
+        self.delay_timer.setInterval(1000)
+        self.delay_timer.timeout.connect(self.on_delay_timer_tick)
+
+        self.setWindowTitle("Cute Sway Recorder")
+        self.config_area = ConfigArea(self)
 
         ## Create labels
-        self.lbl_is_recording = QLabel("Not recording")
+        self.lbl_status = QLabel("Not recording")
 
         ## Create buttons
         self.btn_start_recording = QPushButton("Start recording")
-        self.btn_stop_recording = QPushButton("Stop recording")
+        self.btn_stop_recording = QPushButton(STOP_RECORDING)
 
         ## Connect buttons on-click actions
         self.btn_start_recording.clicked.connect(self.btn_onclick_start_recording)
         self.btn_stop_recording.clicked.connect(self.btn_onclick_stop_recording)
         self.btn_stop_recording.setEnabled(False)
 
-        ## Show window
-        self.window = QWidget()
-        self.config_area = ConfigArea(self.window)
-        self.window.setWindowTitle("Cute Sway Recorder")
-        self.window.setLayout(self.layout())
-        self.window.show()
-
         ## Verify executable dependencies
         self.cmd_available_or_exit("wf-recorder")
         self.cmd_available_or_exit("slurp")
 
+        ## Setup layout
+        self.setup_layout()
+
         ## Define whole-screen icon (not showing yet)
-        self.icon = QSystemTrayIcon(self.window)
-        self.icon.setIcon(self.window.style().standardIcon(QStyle.SP_MediaStop))
+        self.icon = QSystemTrayIcon(self)
+        self.icon.setIcon(self.style().standardIcon(QStyle.SP_MediaStop))
         self.icon.activated.connect(self.tray_icon_activated_handler)
 
-    def layout(self):
+    def on_delay_timer_tick(self):
+        self.delay_time_left -= 1
+        if self.delay_time_left <= 0:
+            self.delay_timer.stop()
+            self.start_recording()
+            self.btn_stop_recording.setText(STOP_RECORDING)
+            return
+        self.lbl_status.setText(STATUS_DELAY.format(self.delay_time_left))
+
+    def setup_layout(self):
         recording_btns = QHBoxLayout()
         recording_btns.addWidget(self.btn_start_recording)
         recording_btns.addWidget(self.btn_stop_recording)
 
         lbl_recording_box = QGroupBox()
         lbl_recording_layout = QVBoxLayout()
-        lbl_recording_layout.addWidget(self.lbl_is_recording, alignment=Qt.AlignCenter)
+        lbl_recording_layout.addWidget(self.lbl_status, alignment=Qt.AlignCenter)
         lbl_recording_box.setLayout(lbl_recording_layout)
 
         layout = QVBoxLayout()
         layout.addWidget(lbl_recording_box)
         layout.addLayout(self.config_area)
         layout.addLayout(recording_btns)
-        return layout
+
+        wid = QWidget()
+        wid.setLayout(layout)
+        self.setCentralWidget(wid)
 
     def cmd_available_or_exit(self, cmd: str):
         """
@@ -117,7 +136,7 @@ class CuteRecorderQtApplication:
                 QMessageBox.Critical,
                 "Executable not found",
                 f"<code>{cmd}</code> not in PATH please install it :(",
-                parent=self.window,
+                parent=self,
             )
             warning.exec()
             sys.exit(1)
@@ -125,43 +144,58 @@ class CuteRecorderQtApplication:
     def tray_icon_activated_handler(self, reason):
         # tray icon was clicked
         if reason == QSystemTrayIcon.Trigger:
-            self.window.show()
+            self.show()
             self.icon.hide()
             self.btn_onclick_stop_recording()
 
     def btn_onclick_start_recording(self):
-        config = self.config_area.create_config()
-        if not config:
+        self.config = self.config_area.create_config()
+        if not self.config:
             return
-
-        # confirm dest file override
-        if config.file_dest.exists():
-            resp = QMessageBox.question(
-                self.window,
-                "File exists",
-                f"Override {config.file_dest}?",
-                QMessageBox.Yes,
-                QMessageBox.No,
-            )
-            if resp == QMessageBox.No:
-                self.lbl_is_recording.setText("Not recording")
-                return
-
-        # show tray icon if recording whole screen
-        if isinstance(config.selection, SelectedScreen):
-            self.window.hide()
-            self.icon.show()
 
         # disable all buttons other than stop record
         self.btn_stop_recording.setEnabled(True)
         self.btn_start_recording.setEnabled(False)
         self.config_area.set_buttons_enabled(False)
 
+        if self.config.delay > 0:
+            self.delay_time_left = self.config.delay
+            self.delay_timer.start()
+            self.lbl_status.setText(STATUS_DELAY.format(self.delay_time_left))
+            self.btn_stop_recording.setText("Cancel countdown")
+        else:
+            self.start_recording()
+
+    def start_recording(self):
+        if not self.config:
+            return
+        conf = self.config
+
+        self.lbl_status.setText(STATUS_RECORDING)
+
+        # confirm dest file override
+        if conf.file_dest.exists():
+            resp = QMessageBox.question(
+                self,
+                "File exists",
+                f"Override {conf.file_dest}?",
+                QMessageBox.Yes,
+                QMessageBox.No,
+            )
+            if resp == QMessageBox.No:
+                self.lbl_status.setText(STATUS_NOT_RECORDING)
+                return
+
+        # show tray icon if recording whole screen
+        if isinstance(conf.selection, SelectedScreen):
+            self.hide()
+            self.icon.show()
+
         # launch wf-recorder
-        self.recorder_proc = start_recording(
-            config.selection,
-            config.file_dest,
-            include_audio=config.include_audio,
+        self.recorder_proc = wf_recorder(
+            conf.selection,
+            conf.file_dest,
+            include_audio=conf.include_audio,
         )
 
     def btn_onclick_stop_recording(self):
@@ -169,19 +203,25 @@ class CuteRecorderQtApplication:
         self.btn_start_recording.setEnabled(True)
         self.config_area.set_buttons_enabled(True)
 
+        if self.delay_timer.isActive():
+            self.delay_timer.stop()
+            self.lbl_status.setText(STATUS_NOT_RECORDING)
+            return
+
         if self.recorder_proc:
             self.recorder_proc.send_signal(signal.SIGINT)
 
-        self.lbl_is_recording.setText('<font color="Green">Saved!</font>')
-
-    def exec(self):
-        return self.app.exec()
+        self.lbl_status.setText('<font color="Green">Saved!</font>')
 
 
 def main():
     subprocess.run(["swaymsg", 'for_window [app_id="cute-sway-recorder"] floating enable'])
-    app = CuteRecorderQtApplication()
-    app.exec()
+    app = QApplication(sys.argv)
+    app.setApplicationDisplayName("Cute Sway Recorder")
+    app.setDesktopFileName("cute-sway-recorder")
+    window = CuteRecorderQtApplication()
+    window.show()
+    sys.exit(app.exec())
 
 
 if __name__ == "__main__":
